+++ apps/common/src/main/scala/com/daml/network/identities/NodeIdentitiesStore.scala
@@ -44,8 +44,7 @@ class NodeIdentitiesStore(
       NodeIdentitiesDump(
         id,
         keys,
-        Some(authorizedStoreSnapshot),
-        None,
+        authorizedStoreSnapshot,
         Some(BuildInfo.compiledVersion),
       )
     }